Ading avatar,header,footer
This commit mailnly deals with the header and footer in the layout page, header rn has the logo and the avatar, the hamburger menu is somthing i am working on, the fotter looks good and the content can be edited by the others. One thing i have in mind for the  header is to give it a transparent bg and have a sticky header so it does not got up with the content on scrolling. Also the search is yet to be implemented. \n So the net commit will contain 3 things: **1. Sticky header 2. Hamburger menu 3. Searchbar/icon**:)